requirements.txt
streamlit==1.29.0
google-generativeai==0.3.1
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
pydub==0.25.1 